+++ packages/core/src/main/conversation/ConversationService.test.node.ts
@@ -17,10 +17,11 @@
  *
  */
 
-import type {CoreCrypto} from '@otak/core-crypto';
+import {CoreCrypto} from '@otak/core-crypto/platforms/web/corecrypto';
 import {APIClient} from '@wireapp/api-client';
 import {ClientType} from '@wireapp/api-client/src/client';
 import {ConversationProtocol} from '@wireapp/api-client/src/conversation';
+import {MlsEvent} from '@wireapp/api-client/src/conversation/data/MlsEventData';
 import {LegalHoldStatus} from '@wireapp/protocol-messaging';
 import {MemoryEngine} from '@wireapp/store-engine';
 import {ConversationService, MessageTargetMode, PayloadBundleSource, PayloadBundleState, PayloadBundleType} from '.';
@@ -32,8 +33,17 @@ import {MessageBuilder} from './message/MessageBuilder';
 import {OtrMessage} from './message/OtrMessage';
 
 describe('ConversationService', () => {
+  beforeAll(() => {
+    jasmine.clock().install();
+  });
+  afterAll(() => {
+    jasmine.clock().uninstall();
+  });
+
   function buildConversationService(federated?: boolean) {
     const client = new APIClient({urls: APIClient.BACKEND.STAGING});
+    spyOn(client.api.conversation, 'postMlsMessage').and.returnValue(Promise.resolve([] as unknown as MlsEvent));
+
     client.context = {
       clientType: ClientType.NONE,
       userId: PayloadHelper.getUUID(),
@@ -45,11 +55,14 @@ describe('ConversationService', () => {
       {
         useQualifiedIds: federated,
       },
-      () => ({} as CoreCrypto),
+      () =>
+        ({
+          encryptMessage: async () => Uint8Array.from([1, 2, 3]),
+        } as unknown as CoreCrypto),
     );
   }
 
-  describe('"send"', () => {
+  describe('"send PROTEUS"', () => {
     const baseMessage = {
       conversation: PayloadHelper.getUUID(),
       from: PayloadHelper.getUUID(),
@@ -71,9 +84,11 @@ describe('ConversationService', () => {
       it(`calls callbacks when sending '${payloadBundle.type}' message is starting and successful`, async () => {
         const conversationService = buildConversationService();
         const sentTime = new Date().toISOString();
-        const onStart = jasmine.createSpy();
+        const onStart = jasmine.createSpy().and.returnValue(Promise.resolve(true));
         const onSuccess = jasmine.createSpy();
+
         spyOn<any>(conversationService, 'sendGenericMessage').and.returnValue(Promise.resolve({time: sentTime}));
+        // const onReconnect = jasmine.createSpy().and.returnValue(getServerAddress());
 
         const promise = conversationService.send({
           protocol: ConversationProtocol.PROTEUS,
@@ -226,8 +241,8 @@ describe('ConversationService', () => {
       const conversationService = buildConversationService();
       spyOn<any>(conversationService, 'sendGenericMessage');
       const message: OtrMessage = {...baseMessage, type: PayloadBundleType.TEXT, content: {text: 'test'}};
-      const onStart = () => Promise.resolve(false);
-      const onSuccess = () => jasmine.createSpy();
+      const onStart = jasmine.createSpy().and.returnValue(Promise.resolve(false));
+      const onSuccess = jasmine.createSpy();
       const payloadBundle = await conversationService.send({
         onStart,
         onSuccess,
@@ -253,6 +268,53 @@ describe('ConversationService', () => {
 
       expect(onSuccess).not.toHaveBeenCalled();
       expect(payloadBundle.state).toBe(PayloadBundleState.CANCELLED);
+    });
+  });
+
+  describe('"send MLS"', () => {
+    const groupId = PayloadHelper.getUUID();
+    const baseMessage = {
+      conversation: PayloadHelper.getUUID(),
+      from: PayloadHelper.getUUID(),
+      id: PayloadHelper.getUUID(),
+      timestamp: 0,
+      source: PayloadBundleSource.LOCAL,
+      state: PayloadBundleState.OUTGOING_UNSENT,
+    };
+    const messages: OtrMessage[] = [
+      {...baseMessage, type: PayloadBundleType.TEXT, content: {text: 'test'}},
+      {
+        ...baseMessage,
+        type: PayloadBundleType.CONFIRMATION,
+        content: {type: 1, firstMessageId: PayloadHelper.getUUID()},
+      },
+      {...baseMessage, type: PayloadBundleType.PING, content: {hotKnock: false}},
+      {
+        ...baseMessage,
+        type: PayloadBundleType.ASSET_IMAGE,
+        content: generateImage(),
+      },
+    ];
+    messages.forEach(payload => {
+      it(`calls callbacks when sending '${payload.type}' message is starting and successful`, async () => {
+        jasmine.clock().mockDate(new Date(0));
+        const conversationService = buildConversationService();
+        const onStart = jasmine.createSpy().and.returnValue(Promise.resolve(true));
+        const onSuccess = jasmine.createSpy();
+
+        const promise = conversationService.send({
+          protocol: ConversationProtocol.MLS,
+          groupId,
+          onStart,
+          onSuccess,
+          payload,
+        });
+
+        expect(onStart).toHaveBeenCalled();
+        expect(onSuccess).not.toHaveBeenCalled();
+        await promise;
+        expect(onSuccess).toHaveBeenCalledWith(jasmine.any(Object), new Date(0).toISOString());
+      });
     });
   });
 
@@ -344,28 +406,13 @@ describe('ConversationService', () => {
 
     it('uploads link previews', async () => {
       const url = 'http://example.com';
-      const image = {
-        data: Buffer.from([]),
-        height: 123,
-        type: 'image/jpeg',
-        width: 456,
-      };
       const text = url;
       const urlOffset = 0;
 
       const linkPreview: LinkPreviewUploadedContent = {
         url,
         urlOffset,
-        imageUploaded: {
-          image,
-          asset: {
-            cipherText: Buffer.from([]),
-            key: '',
-            keyBytes: Buffer.from([]),
-            sha256: Buffer.from([]),
-            token: '',
-          },
-        },
+        imageUploaded: generateImage(),
       };
       const textMessage = MessageBuilder.createText({conversationId: '', from: '', text})
         .withLinkPreviews([linkPreview])
@@ -460,3 +507,22 @@ describe('ConversationService', () => {
     });
   });
 });
+
+function generateImage() {
+  const image = {
+    data: Buffer.from([]),
+    height: 123,
+    type: 'image/jpeg',
+    width: 456,
+  };
+  return {
+    image,
+    asset: {
+      cipherText: Buffer.from([]),
+      key: '',
+      keyBytes: Buffer.from([]),
+      sha256: Buffer.from([]),
+      token: '',
+    },
+  };
+}