+++ packages/changelog-bot/CHANGELOG.md
@@ -3,6 +3,65 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.1.2](https://github.com/wireapp/wire-web-packages/tree/main/packages/changelog-bot/compare/@wireapp/changelog-bot@7.1.1...@wireapp/changelog-bot@7.1.2) (2022-07-08)
+
+**Note:** Version bump only for package @wireapp/changelog-bot
+
+
+
+
+
+## [7.1.1](https://github.com/wireapp/wire-web-packages/tree/main/packages/changelog-bot/compare/@wireapp/changelog-bot@7.1.0...@wireapp/changelog-bot@7.1.1) (2022-07-08)
+
+**Note:** Version bump only for package @wireapp/changelog-bot
+
+
+
+
+
+# [7.1.0](https://github.com/wireapp/wire-web-packages/tree/main/packages/changelog-bot/compare/@wireapp/changelog-bot@7.0.47...@wireapp/changelog-bot@7.1.0) (2022-07-08)
+
+
+### Features
+
+* **core:** Send MLS messages (FS-560) ([#4314](https://github.com/wireapp/wire-web-packages/tree/main/packages/changelog-bot/issues/4314)) ([2f29052](https://github.com/wireapp/wire-web-packages/tree/main/packages/changelog-bot/commit/2f2905233b29c31294ec64e97c1e84998e28821f))
+
+
+
+
+
+## [7.0.47](https://github.com/wireapp/wire-web-packages/tree/main/packages/changelog-bot/compare/@wireapp/changelog-bot@7.0.46...@wireapp/changelog-bot@7.0.47) (2022-07-08)
+
+**Note:** Version bump only for package @wireapp/changelog-bot
+
+
+
+
+
+## [7.0.46](https://github.com/wireapp/wire-web-packages/tree/main/packages/changelog-bot/compare/@wireapp/changelog-bot@7.0.45...@wireapp/changelog-bot@7.0.46) (2022-07-07)
+
+**Note:** Version bump only for package @wireapp/changelog-bot
+
+
+
+
+
+## [7.0.45](https://github.com/wireapp/wire-web-packages/tree/main/packages/changelog-bot/compare/@wireapp/changelog-bot@7.0.44...@wireapp/changelog-bot@7.0.45) (2022-07-07)
+
+**Note:** Version bump only for package @wireapp/changelog-bot
+
+
+
+
+
+## [7.0.44](https://github.com/wireapp/wire-web-packages/tree/main/packages/changelog-bot/compare/@wireapp/changelog-bot@7.0.43...@wireapp/changelog-bot@7.0.44) (2022-07-06)
+
+**Note:** Version bump only for package @wireapp/changelog-bot
+
+
+
+
+
 ## [7.0.43](https://github.com/wireapp/wire-web-packages/tree/main/packages/changelog-bot/compare/@wireapp/changelog-bot@7.0.42...@wireapp/changelog-bot@7.0.43) (2022-07-06)
 
 **Note:** Version bump only for package @wireapp/changelog-bot